Rename TimesliceCircuitBreaker to AdvancedCircuitBreaker
Merge pull request #2 from reisenberger/CircuitBreaker_BranchForRenames
