+++ src/Polly.Shared/AdvancedCircuitBreakerSyntaxAsync.cs
@@ -1,3 +1,5 @@
+﻿#if SUPPORTS_ASYNC
+
 using System;
 using Polly.CircuitBreaker;
 
@@ -6,7 +8,7 @@ namespace Polly
     /// <summary>
     /// Fluent API for defining a Circuit Breaker <see cref="Policy"/>. 
     /// </summary>
-    public static class TimesliceCircuitBreakerSyntax
+    public static class AdvancedCircuitBreakerSyntaxAsync
     {
         /// <summary>
         /// <para> Builds a <see cref="Policy"/> that will function like a Circuit Breaker.</para>
@@ -20,29 +22,28 @@ namespace Polly
         /// </para>
         /// </summary>
         /// <param name="policyBuilder">The policy builder.</param>
-        /// <param name="failureThreshold">The failure threshold at which the circuit will break (a number between 0 and 1; eg 0.5 represents breaking if 50% or more of actions result in a handled failure).</param>
+        /// <param name="failureThreshold">The failure threshold at which the circuit will break (a number between 0 and 1; eg 0.5 represents breaking if 50% or more of actions result in a handled failure.</param>
         /// <param name="timesliceDuration">The duration of the timeslice over which failure ratios are assessed.</param>
         /// <param name="minimumThroughput">The minimum throughput: this many actions or more must pass through the circuit in the timeslice, for statistics to be considered significant and the circuit-breaker to come into action.</param>
         /// <param name="durationOfBreak">The duration the circuit will stay open before resetting.</param>
         /// <returns>The policy instance.</returns>
         /// <remarks>(see "Release It!" by Michael T. Nygard fi)</remarks>
         /// <exception cref="System.ArgumentOutOfRangeException">exceptionsAllowedBeforeBreaking;Value must be greater than zero.</exception>
-        /// <exception cref="ArgumentNullException">onBreak</exception>
-        /// <exception cref="ArgumentNullException">onReset</exception>
-        public static CircuitBreakerPolicy TimesliceCircuitBreaker(this PolicyBuilder policyBuilder, double failureThreshold, TimeSpan timesliceDuration, int minimumThroughput, TimeSpan durationOfBreak)
+        public static CircuitBreakerPolicy AdvancedCircuitBreakerAsync(this PolicyBuilder policyBuilder, double failureThreshold, TimeSpan timesliceDuration, int minimumThroughput, TimeSpan durationOfBreak)
         {
             Action<Exception, TimeSpan> doNothingOnBreak = (_, __) => { };
             Action doNothingOnReset = () => { };
 
-            return policyBuilder.TimesliceCircuitBreaker(
-                failureThreshold, timesliceDuration, minimumThroughput, 
-                durationOfBreak, 
-                doNothingOnBreak,
-                doNothingOnReset
-                );
+            return policyBuilder.AdvancedCircuitBreakerAsync(
+               failureThreshold, timesliceDuration, minimumThroughput, 
+               durationOfBreak,
+               doNothingOnBreak,
+               doNothingOnReset
+               );
         }
 
         /// <summary>
+        /// <para> Builds a <see cref="Policy"/> that will function like a Circuit Breaker.</para>
         /// <para>The circuit will break if, within any timeslice of duration <paramref name="timesliceDuration"/>, the proportion of actions resulting in a handled exception exceeds <paramref name="failureThreshold"/>, provided also that the number of actions through the circuit in the timeslice is at least <paramref name="minimumThroughput" />. </para>
         /// <para>The circuit will stay broken for the <paramref name="durationOfBreak" />. Any attempt to execute this policy
         /// while the circuit is broken, will immediately throw a <see cref="BrokenCircuitException" /> containing the exception
@@ -64,17 +65,18 @@ namespace Polly
         /// <exception cref="System.ArgumentOutOfRangeException">exceptionsAllowedBeforeBreaking;Value must be greater than zero.</exception>
         /// <exception cref="ArgumentNullException">onBreak</exception>
         /// <exception cref="ArgumentNullException">onReset</exception>
-        public static CircuitBreakerPolicy TimesliceCircuitBreaker(this PolicyBuilder policyBuilder, double failureThreshold, TimeSpan timesliceDuration, int minimumThroughput, TimeSpan durationOfBreak, Action<Exception, TimeSpan> onBreak, Action onReset)
+        public static CircuitBreakerPolicy AdvancedCircuitBreakerAsync(this PolicyBuilder policyBuilder, double failureThreshold, TimeSpan timesliceDuration, int minimumThroughput, TimeSpan durationOfBreak, Action<Exception, TimeSpan> onBreak, Action onReset)
         {
-            return policyBuilder.TimesliceCircuitBreaker(
+            return policyBuilder.AdvancedCircuitBreakerAsync(
                 failureThreshold, timesliceDuration, minimumThroughput, 
-                durationOfBreak, 
-                (exception, timespan, context) => onBreak(exception, timespan), 
+                durationOfBreak,
+                (exception, timespan, context) => onBreak(exception, timespan),
                 context => onReset()
                 );
         }
 
         /// <summary>
+        /// <para> Builds a <see cref="Policy"/> that will function like a Circuit Breaker.</para>
         /// <para>The circuit will break if, within any timeslice of duration <paramref name="timesliceDuration"/>, the proportion of actions resulting in a handled exception exceeds <paramref name="failureThreshold"/>, provided also that the number of actions through the circuit in the timeslice is at least <paramref name="minimumThroughput" />. </para>
         /// <para>The circuit will stay broken for the <paramref name="durationOfBreak" />. Any attempt to execute this policy
         /// while the circuit is broken, will immediately throw a <see cref="BrokenCircuitException" /> containing the exception
@@ -96,10 +98,11 @@ namespace Polly
         /// <exception cref="System.ArgumentOutOfRangeException">exceptionsAllowedBeforeBreaking;Value must be greater than zero.</exception>
         /// <exception cref="ArgumentNullException">onBreak</exception>
         /// <exception cref="ArgumentNullException">onReset</exception>
-        public static CircuitBreakerPolicy TimesliceCircuitBreaker(this PolicyBuilder policyBuilder, double failureThreshold, TimeSpan timesliceDuration, int minimumThroughput, TimeSpan durationOfBreak, Action<Exception, TimeSpan, Context> onBreak, Action<Context> onReset)
+        public static CircuitBreakerPolicy AdvancedCircuitBreakerAsync(this PolicyBuilder policyBuilder, double failureThreshold, TimeSpan timesliceDuration, int minimumThroughput, TimeSpan durationOfBreak, Action<Exception, TimeSpan, Context> onBreak, Action<Context> onReset)
         {
             Action doNothingOnHalfOpen = () => { };
-            return policyBuilder.TimesliceCircuitBreaker(failureThreshold, timesliceDuration, minimumThroughput, 
+            return policyBuilder.AdvancedCircuitBreakerAsync(
+                failureThreshold, timesliceDuration, minimumThroughput, 
                 durationOfBreak, 
                 onBreak, 
                 onReset,
@@ -108,6 +111,7 @@ namespace Polly
         }
 
         /// <summary>
+        /// <para> Builds a <see cref="Policy"/> that will function like a Circuit Breaker.</para>
         /// <para>The circuit will break if, within any timeslice of duration <paramref name="timesliceDuration"/>, the proportion of actions resulting in a handled exception exceeds <paramref name="failureThreshold"/>, provided also that the number of actions through the circuit in the timeslice is at least <paramref name="minimumThroughput" />. </para>
         /// <para>The circuit will stay broken for the <paramref name="durationOfBreak" />. Any attempt to execute this policy
         /// while the circuit is broken, will immediately throw a <see cref="BrokenCircuitException" /> containing the exception
@@ -130,9 +134,9 @@ namespace Polly
         /// <exception cref="System.ArgumentOutOfRangeException">exceptionsAllowedBeforeBreaking;Value must be greater than zero.</exception>
         /// <exception cref="ArgumentNullException">onBreak</exception>
         /// <exception cref="ArgumentNullException">onReset</exception>
-        public static CircuitBreakerPolicy TimesliceCircuitBreaker(this PolicyBuilder policyBuilder, double failureThreshold, TimeSpan timesliceDuration, int minimumThroughput, TimeSpan durationOfBreak, Action<Exception, TimeSpan> onBreak, Action onReset, Action onHalfOpen)
+        public static CircuitBreakerPolicy AdvancedCircuitBreakerAsync(this PolicyBuilder policyBuilder, double failureThreshold, TimeSpan timesliceDuration, int minimumThroughput, TimeSpan durationOfBreak, Action<Exception, TimeSpan> onBreak, Action onReset, Action onHalfOpen)
         {
-            return policyBuilder.TimesliceCircuitBreaker(
+            return policyBuilder.AdvancedCircuitBreakerAsync(
                 failureThreshold, timesliceDuration, minimumThroughput, 
                 durationOfBreak,
                 (exception, timespan, context) => onBreak(exception, timespan),
@@ -142,7 +146,7 @@ namespace Polly
         }
 
         /// <summary>
-        /// <para> Builds a <see cref="Policy" /> that will function like a Circuit Breaker.</para>
+        /// <para> Builds a <see cref="Policy"/> that will function like a Circuit Breaker.</para>
         /// <para>The circuit will break if, within any timeslice of duration <paramref name="timesliceDuration"/>, the proportion of actions resulting in a handled exception exceeds <paramref name="failureThreshold"/>, provided also that the number of actions through the circuit in the timeslice is at least <paramref name="minimumThroughput" />. </para>
         /// <para>The circuit will stay broken for the <paramref name="durationOfBreak" />. Any attempt to execute this policy
         /// while the circuit is broken, will immediately throw a <see cref="BrokenCircuitException" /> containing the exception
@@ -157,25 +161,18 @@ namespace Polly
         /// <param name="timesliceDuration">The duration of the timeslice over which failure ratios are assessed.</param>
         /// <param name="minimumThroughput">The minimum throughput: this many actions or more must pass through the circuit in the timeslice, for statistics to be considered significant and the circuit-breaker to come into action.</param>
         /// <param name="durationOfBreak">The duration the circuit will stay open before resetting.</param>
-        /// <param name="onBreak">The action to call when the circuit transitions to an <see cref="CircuitState.Open" /> state.</param>
-        /// <param name="onReset">The action to call when the circuit resets to a <see cref="CircuitState.Closed" /> state.</param>
-        /// <param name="onHalfOpen">The action to call when the circuit transitions to <see cref="CircuitState.HalfOpen" /> state, ready to try action executions again.</param>
+        /// <param name="onBreak">The action to call when the circuit transitions to an <see cref="CircuitState.Open"/> state.</param>
+        /// <param name="onReset">The action to call when the circuit resets to a <see cref="CircuitState.Closed"/> state.</param>
+        /// <param name="onHalfOpen">The action to call when the circuit transitions to <see cref="CircuitState.HalfOpen"/> state, ready to try action executions again. </param>
         /// <returns>The policy instance.</returns>
+        /// <remarks>(see "Release It!" by Michael T. Nygard fi)</remarks>
         /// <exception cref="System.ArgumentOutOfRangeException">exceptionsAllowedBeforeBreaking;Value must be greater than zero.</exception>
-        /// <exception cref="System.ArgumentNullException">
-        /// onBreak
-        /// or
-        /// onReset
-        /// or
-        /// onHalfOpen
-        /// </exception>
-        /// <exception cref="ArgumentNullException">exceptionsAllowedBeforeBreaking;Value must be greater than zero.</exception>
         /// <exception cref="ArgumentNullException">onBreak</exception>
         /// <exception cref="ArgumentNullException">onReset</exception>
-        /// <remarks>(see "Release It!" by Michael T. Nygard fi)</remarks>
-        public static CircuitBreakerPolicy TimesliceCircuitBreaker(this PolicyBuilder policyBuilder, double failureThreshold, TimeSpan timesliceDuration, int minimumThroughput, TimeSpan durationOfBreak, Action<Exception, TimeSpan, Context> onBreak, Action<Context> onReset, Action onHalfOpen)
+        /// <exception cref="ArgumentNullException">onHalfOpen</exception>
+        public static CircuitBreakerPolicy AdvancedCircuitBreakerAsync(this PolicyBuilder policyBuilder, double failureThreshold, TimeSpan timesliceDuration, int minimumThroughput, TimeSpan durationOfBreak, Action<Exception, TimeSpan, Context> onBreak, Action<Context> onReset, Action onHalfOpen)
         {
-            var resolutionOfCircuit = TimeSpan.FromTicks(TimesliceCircuitController.ResolutionOfCircuitTimer);
+            var resolutionOfCircuit = TimeSpan.FromTicks(AdvancedCircuitController.ResolutionOfCircuitTimer);
 
             if (failureThreshold <= 0) throw new ArgumentOutOfRangeException("failureThreshold", "Value must be greater than zero.");
             if (failureThreshold > 1) throw new ArgumentOutOfRangeException("failureThreshold", "Value must be less than or equal to one.");
@@ -187,12 +184,14 @@ namespace Polly
             if (onReset == null) throw new ArgumentNullException("onReset");
             if (onHalfOpen == null) throw new ArgumentNullException("onHalfOpen");
 
-            var breakerController = new TimesliceCircuitController(failureThreshold, timesliceDuration, minimumThroughput, durationOfBreak, onBreak, onReset, onHalfOpen);
+            var breakerController = new AdvancedCircuitController(failureThreshold, timesliceDuration, minimumThroughput, durationOfBreak, onBreak, onReset, onHalfOpen);
             return new CircuitBreakerPolicy(
-                (action, context) => CircuitBreakerEngine.Implementation(action, context, policyBuilder.ExceptionPredicates, breakerController), 
-                policyBuilder.ExceptionPredicates, 
+                (action, context, cancellationToken, continueOnCapturedContext) => CircuitBreakerEngine.ImplementationAsync(action, context, policyBuilder.ExceptionPredicates, breakerController, cancellationToken, continueOnCapturedContext),
+                policyBuilder.ExceptionPredicates,
                 breakerController
-                );
+            );
         }
     }
 }
+
+#endif
