+++ src/Polly.SharedSpecs/AdvancedCircuitBreakerSpecs.cs
@@ -1,18 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Threading;
-using System.Threading.Tasks;
 using FluentAssertions;
 using Polly.CircuitBreaker;
 using Polly.Specs.Helpers;
 using Polly.Utilities;
 using Xunit;
 
-using Scenario = Polly.Specs.Helpers.PolicyExtensions.ExceptionAndOrCancellationScenario;
-
 namespace Polly.Specs
 {
-    public class TimesliceCircuitBreakerAsyncSpecs : IDisposable
+    public class AdvancedCircuitBreakerSpecs : IDisposable
     {
         #region Configuration tests
 
@@ -21,9 +17,9 @@ namespace Polly.Specs
         {
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), 4, TimeSpan.MaxValue);
+                .AdvancedCircuitBreaker(0.5, TimeSpan.FromSeconds(10), 4, TimeSpan.MaxValue);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
         }
 
@@ -32,7 +28,7 @@ namespace Polly.Specs
         {
             Action action = () => Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0, TimeSpan.FromSeconds(10), 4, TimeSpan.FromSeconds(30));
+                .AdvancedCircuitBreaker(0, TimeSpan.FromSeconds(10), 4, TimeSpan.FromSeconds(30));
 
             action.ShouldThrow<ArgumentOutOfRangeException>()
                 .And.ParamName.Should()
@@ -44,7 +40,7 @@ namespace Polly.Specs
         {
             Action action = () => Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(-0.5, TimeSpan.FromSeconds(10), 4, TimeSpan.FromSeconds(30));
+                .AdvancedCircuitBreaker(-0.5, TimeSpan.FromSeconds(10), 4, TimeSpan.FromSeconds(30));
 
             action.ShouldThrow<ArgumentOutOfRangeException>()
                 .And.ParamName.Should()
@@ -56,7 +52,7 @@ namespace Polly.Specs
         {
             Action action = () => Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(1.0, TimeSpan.FromSeconds(10), 4, TimeSpan.FromSeconds(30));
+                .AdvancedCircuitBreaker(1.0, TimeSpan.FromSeconds(10), 4, TimeSpan.FromSeconds(30));
 
             action.ShouldNotThrow();
         }
@@ -66,7 +62,7 @@ namespace Polly.Specs
         {
             Action action = () => Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(1.01, TimeSpan.FromSeconds(10), 4, TimeSpan.FromSeconds(30));
+                .AdvancedCircuitBreaker(1.01, TimeSpan.FromSeconds(10), 4, TimeSpan.FromSeconds(30));
 
             action.ShouldThrow<ArgumentOutOfRangeException>()
                 .And.ParamName.Should()
@@ -78,7 +74,7 @@ namespace Polly.Specs
         {
             Action action = () => Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     0.5, 
                     TimeSpan.FromMilliseconds(20).Add(TimeSpan.FromTicks(-1)), 
                     4, 
@@ -94,7 +90,7 @@ namespace Polly.Specs
         {
             Action action = () => Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromMilliseconds(20), 4, TimeSpan.FromSeconds(30));
+                .AdvancedCircuitBreaker(0.5, TimeSpan.FromMilliseconds(20), 4, TimeSpan.FromSeconds(30));
 
             action.ShouldNotThrow<ArgumentOutOfRangeException>();
         }
@@ -104,7 +100,7 @@ namespace Polly.Specs
         {
             Action action = () => Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), 0, TimeSpan.FromSeconds(30));
+                .AdvancedCircuitBreaker(0.5, TimeSpan.FromSeconds(10), 0, TimeSpan.FromSeconds(30));
 
             action.ShouldThrow<ArgumentOutOfRangeException>()
                 .And.ParamName.Should()
@@ -116,7 +112,7 @@ namespace Polly.Specs
         {
             Action action = () => Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), -1, TimeSpan.FromSeconds(30));
+                .AdvancedCircuitBreaker(0.5, TimeSpan.FromSeconds(10), -1, TimeSpan.FromSeconds(30));
 
             action.ShouldThrow<ArgumentOutOfRangeException>()
                 .And.ParamName.Should()
@@ -128,7 +124,7 @@ namespace Polly.Specs
         {
             Action action = () => Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), 4, -TimeSpan.FromSeconds(1));
+                .AdvancedCircuitBreaker(0.5, TimeSpan.FromSeconds(10), 4, -TimeSpan.FromSeconds(1));
 
             action.ShouldThrow<ArgumentOutOfRangeException>()
                 .And.ParamName.Should()
@@ -140,7 +136,7 @@ namespace Polly.Specs
         {
             Action action = () => Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), 4, TimeSpan.Zero);
+                .AdvancedCircuitBreaker(0.5, TimeSpan.FromSeconds(10), 4, TimeSpan.Zero);
 
             action.ShouldNotThrow();
         }
@@ -150,7 +146,7 @@ namespace Polly.Specs
         {
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), 4, TimeSpan.FromSeconds(30));
+                .AdvancedCircuitBreaker(0.5, TimeSpan.FromSeconds(10), 4, TimeSpan.FromSeconds(30));
 
             breaker.CircuitState.Should().Be(CircuitState.Closed);
         }
@@ -161,7 +157,7 @@ namespace Polly.Specs
 
         #region Tests that are independent from health metrics implementation
 
-        // Tests on the TimesliceCircuitBreakerAsync operation typically use a breaker: 
+        // Tests on the AdvancedCircuitBreaker operation typically use a breaker: 
         // - with a failure threshold of >=50%, 
         // - and a throughput threshold of 4
         // - across a ten-second period.
@@ -175,7 +171,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromSeconds(10),
                     minimumThroughput: 4,
@@ -183,21 +179,21 @@ namespace Polly.Specs
                 );
 
             // Three of three actions in this test throw handled failures.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
             // Failure threshold exceeded, but throughput threshold not yet.
 
             // Throughput threshold will be exceeded by the below successful call, but we never break on a successful call; hence don't break on this.
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
             // No adjustment to SystemClock.UtcNow, so all exceptions were raised within same timeslice
@@ -212,7 +208,7 @@ namespace Polly.Specs
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
                 .Or<ArgumentOutOfRangeException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromSeconds(10),
                     minimumThroughput: 4,
@@ -220,19 +216,19 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw unhandled failures.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<ArgumentNullException>())
+            breaker.Invoking(x => x.RaiseException<ArgumentNullException>())
                 .ShouldThrow<ArgumentNullException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<ArgumentNullException>())
+            breaker.Invoking(x => x.RaiseException<ArgumentNullException>())
                 .ShouldThrow<ArgumentNullException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<ArgumentNullException>())
+            breaker.Invoking(x => x.RaiseException<ArgumentNullException>())
                 .ShouldThrow<ArgumentNullException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<ArgumentNullException>())
+            breaker.Invoking(x => x.RaiseException<ArgumentNullException>())
                 .ShouldThrow<ArgumentNullException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
         }
@@ -241,7 +237,7 @@ namespace Polly.Specs
 
         #region With sample duration higher than 199 ms so that multiple windows are used
 
-        // Tests on the TimesliceCircuitBreakerAsync operation typically use a breaker: 
+        // Tests on the AdvancedCircuitBreaker operation typically use a breaker: 
         // - with a failure threshold of >=50%, 
         // - and a throughput threshold of 4
         // - across a ten-second period.
@@ -255,7 +251,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromSeconds(10),
                     minimumThroughput: 4,
@@ -263,25 +259,25 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
             // No adjustment to SystemClock.UtcNow, so all exceptions raised within same timeslice
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<BrokenCircuitException>()
                 .WithMessage("The circuit is now open and is not allowing calls.")
                 .WithInnerException<DivideByZeroException>();
@@ -299,7 +295,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: timesliceDuration,
                     minimumThroughput: 4,
@@ -307,15 +303,15 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
@@ -324,11 +320,11 @@ namespace Polly.Specs
             // They are still placed within same timeslice.
             SystemClock.UtcNow = () => time.AddSeconds(timesliceDuration.Seconds / 2d + 1);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<BrokenCircuitException>()
                 .WithMessage("The circuit is now open and is not allowing calls.")
                 .WithInnerException<DivideByZeroException>();
@@ -344,7 +340,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromSeconds(10),
                     minimumThroughput: 4,
@@ -352,24 +348,24 @@ namespace Polly.Specs
                 );
 
             // Three of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
             // No adjustment to SystemClock.UtcNow, so all exceptions were raised within same timeslice
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<BrokenCircuitException>()
                 .WithMessage("The circuit is now open and is not allowing calls.")
                 .WithInnerException<DivideByZeroException>();
@@ -387,7 +383,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: timesliceDuration,
                     minimumThroughput: 4,
@@ -395,19 +391,19 @@ namespace Polly.Specs
                 );
 
             // Three of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
 
@@ -416,7 +412,7 @@ namespace Polly.Specs
             // They are still placed within same timeslice
             SystemClock.UtcNow = () => time.AddSeconds(timesliceDuration.Seconds / 2d + 1);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<BrokenCircuitException>()
                 .WithMessage("The circuit is now open and is not allowing calls.")
                 .WithInnerException<DivideByZeroException>();
@@ -432,7 +428,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromSeconds(10),
                     minimumThroughput: 4,
@@ -440,24 +436,24 @@ namespace Polly.Specs
                 );
 
             // Two of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
             // No adjustment to SystemClock.UtcNow, so all exceptions were raised within same timeslice
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<BrokenCircuitException>()
                 .WithMessage("The circuit is now open and is not allowing calls.")
                 .WithInnerException<DivideByZeroException>();
@@ -475,7 +471,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: timesliceDuration,
                     minimumThroughput: 4,
@@ -483,19 +479,19 @@ namespace Polly.Specs
                 );
 
             // Two of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
 
@@ -504,7 +500,7 @@ namespace Polly.Specs
             // They are still placed within same timeslice
             SystemClock.UtcNow = () => time.AddSeconds(timesliceDuration.Seconds / 2d + 1);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<BrokenCircuitException>()
                 .WithMessage("The circuit is now open and is not allowing calls.")
                 .WithInnerException<DivideByZeroException>();
@@ -522,7 +518,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: timesliceDuration,
                     minimumThroughput: 4,
@@ -530,22 +526,22 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures; but only the first three within the timeslice.
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
             // Adjust SystemClock so that timeslice (clearly) expires; fourth exception thrown in next-recorded timeslice.
             SystemClock.UtcNow = () => time.Add(timesliceDuration).Add(timesliceDuration);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
@@ -561,7 +557,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: timesliceDuration,
                     minimumThroughput: 4,
@@ -569,22 +565,22 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures; but only the first three within the timeslice.
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
             // Adjust SystemClock so that timeslice (just) expires; fourth exception thrown in following timeslice.
             SystemClock.UtcNow = () => time.Add(timesliceDuration);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
@@ -600,7 +596,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: timesliceDuration,
                     minimumThroughput: 4,
@@ -610,25 +606,25 @@ namespace Polly.Specs
             // Four of four actions in this test throw handled failures; but only the first three within the original timeslice.
 
             // Two actions at the start of the original timeslice.
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
             // Creates a new window right at the end of the original timeslice.
             SystemClock.UtcNow = () => time.AddTicks(timesliceDuration.Ticks - 1);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
             // Adjust SystemClock so that timeslice (just) expires; fourth exception thrown in following timeslice.  If timeslice/window rollover is precisely defined, this should cause first two actions to be forgotten from statistics (rolled out of the window of relevance), and thus the circuit not to break.
             SystemClock.UtcNow = () => time.Add(timesliceDuration);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
@@ -644,7 +640,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: timesliceDuration,
                     minimumThroughput: 4,
@@ -652,26 +648,26 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
             // Adjust SystemClock so that timeslice doesn't quite expire; fourth exception thrown in same timeslice.
             SystemClock.UtcNow = () => time.AddTicks(timesliceDuration.Ticks - 1);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<BrokenCircuitException>()
                 .WithMessage("The circuit is now open and is not allowing calls.")
                 .WithInnerException<DivideByZeroException>();
@@ -687,7 +683,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromSeconds(10),
                     minimumThroughput: 4,
@@ -695,15 +691,15 @@ namespace Polly.Specs
                 );
 
             // One of three actions in this test throw handled failures.
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
             // No adjustment to SystemClock.UtcNow, so all exceptions were raised within same timeslice
@@ -717,7 +713,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromSeconds(10),
                     minimumThroughput: 4,
@@ -725,19 +721,19 @@ namespace Polly.Specs
                 );
 
             // One of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
             // No adjustment to SystemClock.UtcNow, so all exceptions were raised within same timeslice
@@ -753,7 +749,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: timesliceDuration,
                     minimumThroughput: 4,
@@ -762,7 +758,7 @@ namespace Polly.Specs
 
             // Executing a single invocation to ensure timeslice is created
             // This invocation is not be counted against the threshold
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
@@ -771,15 +767,15 @@ namespace Polly.Specs
             SystemClock.UtcNow = () => time.AddTicks(timesliceDuration.Ticks - 1);
 
             // Three of four actions in this test occur within the first timeslice.
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
@@ -789,7 +785,7 @@ namespace Polly.Specs
             // This failure opens the circuit, because it is the second failure of four calls
             // equalling the failure threshold. The minimum threshold within the defined
             // timeslice duration is met, when using rolling windows.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
         }
@@ -804,7 +800,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: timesliceDuration,
                     minimumThroughput: 4,
@@ -813,7 +809,7 @@ namespace Polly.Specs
 
             // Executing a single invocation to ensure timeslice is created
             // This invocation is not be counted against the threshold
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
@@ -822,11 +818,11 @@ namespace Polly.Specs
             SystemClock.UtcNow = () => time.AddTicks(timesliceDuration.Ticks - 1);
 
             // Two of three actions in this test occur within the first timeslice.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
@@ -836,7 +832,7 @@ namespace Polly.Specs
             // A third failure occurs just at the beginning of the new timeslice making 
             // the number of failures above the failure threshold. However, the throughput is 
             // below the minimum threshold as to open the circuit.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
         }
@@ -852,7 +848,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: timesliceDuration,
                     minimumThroughput: 4,
@@ -861,7 +857,7 @@ namespace Polly.Specs
 
             // Executing a single invocation to ensure timeslice is created
             // This invocation is not be counted against the threshold
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
@@ -869,22 +865,22 @@ namespace Polly.Specs
             SystemClock.UtcNow = () => time.AddSeconds(timesliceDuration.Seconds / (double)numberOfWindowsDefinedInCircuitBreaker);
 
             // Three actions occur in the second window of the first timeslice
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
             // Setting the time to just barely into the new timeslice
             SystemClock.UtcNow = () => time.Add(timesliceDuration);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
         }
@@ -893,7 +889,7 @@ namespace Polly.Specs
 
         #region With sample duration at 199 ms so that only a single window is used
 
-        // These tests on TimesliceCircuitBreakerAsync operation typically use a breaker: 
+        // These tests on AdvancedCircuitBreaker operation typically use a breaker: 
         // - with a failure threshold of >=50%, 
         // - and a throughput threshold of 4
         // - across a 199ms period.
@@ -907,7 +903,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromMilliseconds(199),
                     minimumThroughput: 4,
@@ -915,25 +911,25 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
             // No adjustment to SystemClock.UtcNow, so all exceptions raised within same timeslice
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<BrokenCircuitException>()
                 .WithMessage("The circuit is now open and is not allowing calls.")
                 .WithInnerException<DivideByZeroException>();
@@ -949,7 +945,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromMilliseconds(199),
                     minimumThroughput: 4,
@@ -957,24 +953,24 @@ namespace Polly.Specs
                 );
 
             // Three of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
             // No adjustment to SystemClock.UtcNow, so all exceptions were raised within same timeslice
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<BrokenCircuitException>()
                 .WithMessage("The circuit is now open and is not allowing calls.")
                 .WithInnerException<DivideByZeroException>();
@@ -990,7 +986,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromMilliseconds(199),
                     minimumThroughput: 4,
@@ -998,24 +994,24 @@ namespace Polly.Specs
                 );
 
             // Two of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
             // No adjustment to SystemClock.UtcNow, so all exceptions were raised within same timeslice
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<BrokenCircuitException>()
                 .WithMessage("The circuit is now open and is not allowing calls.")
                 .WithInnerException<DivideByZeroException>();
@@ -1033,7 +1029,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: timesliceDuration,
                     minimumThroughput: 4,
@@ -1041,22 +1037,22 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures; but only the first within the timeslice.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
             // Adjust SystemClock so that timeslice (clearly) expires; fourth exception thrown in next-recorded timeslice.
             SystemClock.UtcNow = () => time.Add(timesliceDuration).Add(timesliceDuration);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
@@ -1072,7 +1068,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: timesliceDuration,
                     minimumThroughput: 4,
@@ -1080,22 +1076,22 @@ namespace Polly.Specs
                 );
 
             // Two of four actions in this test throw handled failures; but only the first within the timeslice.
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
             // Adjust SystemClock so that timeslice (just) expires; fourth exception thrown in following timeslice.
             SystemClock.UtcNow = () => time.Add(timesliceDuration);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
@@ -1111,7 +1107,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: timesliceDuration,
                     minimumThroughput: 4,
@@ -1119,26 +1115,26 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
             // Adjust SystemClock so that timeslice doesn't quite expire; fourth exception thrown in same timeslice.
             SystemClock.UtcNow = () => time.AddTicks(timesliceDuration.Ticks - 1);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<BrokenCircuitException>()
                 .WithMessage("The circuit is now open and is not allowing calls.")
                 .WithInnerException<DivideByZeroException>();
@@ -1154,7 +1150,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromMilliseconds(199),
                     minimumThroughput: 4,
@@ -1162,15 +1158,15 @@ namespace Polly.Specs
                 );
 
             // One of three actions in this test throw handled failures.
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
             // No adjustment to SystemClock.UtcNow, so all exceptions were raised within same timeslice
@@ -1184,7 +1180,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromMilliseconds(199),
                     minimumThroughput: 4,
@@ -1192,19 +1188,19 @@ namespace Polly.Specs
                 );
 
             // One of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
             // No adjustment to SystemClock.UtcNow, so all exceptions were raised within same timeslice
@@ -1220,7 +1216,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: timesliceDuration,
                     minimumThroughput: 4,
@@ -1229,7 +1225,7 @@ namespace Polly.Specs
 
             // Executing a single invocation to ensure timeslice is created
             // This invocation is not be counted against the threshold
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
@@ -1238,15 +1234,15 @@ namespace Polly.Specs
             SystemClock.UtcNow = () => time.AddTicks(timesliceDuration.Ticks - 1);
 
             // Three of four actions in this test occur within the first timeslice.
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldNotThrow();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
@@ -1255,7 +1251,7 @@ namespace Polly.Specs
 
             // This failure does not open the circuit, because a new duration should have 
             // started and with such low sampling duration, windows should not be used.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
         }
@@ -1276,7 +1272,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromSeconds(10),
                     minimumThroughput: 4,
@@ -1284,19 +1280,19 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
 
             breaker.CircuitState.Should().Be(CircuitState.Open);
@@ -1305,7 +1301,7 @@ namespace Polly.Specs
 
             // duration has passed, circuit now half open
             breaker.CircuitState.Should().Be(CircuitState.HalfOpen);
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
         }
 
@@ -1320,7 +1316,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: timesliceDuration,
                     minimumThroughput: 4,
@@ -1328,15 +1324,15 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
@@ -1346,7 +1342,7 @@ namespace Polly.Specs
             var anotherwindowDuration = timesliceDuration.Seconds / 2d + 1;
             SystemClock.UtcNow = () => time.AddSeconds(anotherwindowDuration);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
 
             breaker.CircuitState.Should().Be(CircuitState.Open);
@@ -1360,7 +1356,7 @@ namespace Polly.Specs
 
             // duration has passed, circuit now half open
             breaker.CircuitState.Should().Be(CircuitState.HalfOpen);
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
         }
 
@@ -1374,7 +1370,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromSeconds(10),
                     minimumThroughput: 4,
@@ -1382,19 +1378,19 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
 
             breaker.CircuitState.Should().Be(CircuitState.Open);
@@ -1405,16 +1401,16 @@ namespace Polly.Specs
             breaker.CircuitState.Should().Be(CircuitState.HalfOpen);
 
             // first call after duration raises an exception, so circuit should open again
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<BrokenCircuitException>();
-
+            
         }
 
         [Fact]
-        public async Task Should_reset_circuit_after_the_specified_duration_has_passed_if_the_next_call_does_not_raise_an_exception()
+        public void Should_reset_circuit_after_the_specified_duration_has_passed_if_the_next_call_does_not_raise_an_exception()
         {
             var time = 1.January(2000);
             SystemClock.UtcNow = () => time;
@@ -1423,7 +1419,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromSeconds(10),
                     minimumThroughput: 4,
@@ -1431,19 +1427,19 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
 
             breaker.CircuitState.Should().Be(CircuitState.Open);
@@ -1454,7 +1450,7 @@ namespace Polly.Specs
             breaker.CircuitState.Should().Be(CircuitState.HalfOpen);
 
             // first call after duration is successful, so circuit should reset
-            await breaker.ExecuteAsync(() => Task.FromResult(true));
+            breaker.Execute(() => {});
             breaker.CircuitState.Should().Be(CircuitState.Closed);
         }
 
@@ -1467,7 +1463,7 @@ namespace Polly.Specs
         {
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), 4, TimeSpan.FromSeconds(30));
+                .AdvancedCircuitBreaker(0.5, TimeSpan.FromSeconds(10), 4, TimeSpan.FromSeconds(30));
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
             // manually break circuit
@@ -1475,7 +1471,7 @@ namespace Polly.Specs
             breaker.CircuitState.Should().Be(CircuitState.Isolated);
 
             // circuit manually broken: execution should be blocked; even non-exception-throwing executions should not reset circuit
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldThrow<IsolatedCircuitException>();
             breaker.CircuitState.Should().Be(CircuitState.Isolated);
         }
@@ -1489,7 +1485,7 @@ namespace Polly.Specs
             var durationOfBreak = TimeSpan.FromSeconds(30);
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), 4, durationOfBreak);
+                .AdvancedCircuitBreaker(0.5, TimeSpan.FromSeconds(10), 4, durationOfBreak);
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
             breaker.Isolate();
@@ -1497,7 +1493,7 @@ namespace Polly.Specs
 
             SystemClock.UtcNow = () => time.Add(durationOfBreak);
             breaker.CircuitState.Should().Be(CircuitState.Isolated);
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldThrow<IsolatedCircuitException>();
         }
 
@@ -1510,18 +1506,17 @@ namespace Polly.Specs
             var durationOfBreak = TimeSpan.FromSeconds(30);
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), 4, durationOfBreak);
+                .AdvancedCircuitBreaker(0.5, TimeSpan.FromSeconds(10), 4, durationOfBreak);
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
             breaker.Isolate();
             breaker.CircuitState.Should().Be(CircuitState.Isolated);
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldThrow<IsolatedCircuitException>();
 
             breaker.Reset();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
-                .ShouldNotThrow();
+            breaker.Invoking(x => x.Execute(() => { })).ShouldNotThrow();
         }
 
         [Fact]
@@ -1534,7 +1529,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromSeconds(10),
                     minimumThroughput: 4,
@@ -1542,19 +1537,19 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
 
@@ -1562,8 +1557,7 @@ namespace Polly.Specs
             breaker.Reset();
             SystemClock.UtcNow().Should().Be(time);
             breaker.CircuitState.Should().Be(CircuitState.Closed);
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
-                .ShouldNotThrow();
+            breaker.Invoking(x => x.Execute(() => { })).ShouldNotThrow();
         }
 
         #endregion
@@ -1580,7 +1574,7 @@ namespace Polly.Specs
             var durationOfBreak = TimeSpan.FromSeconds(30);
             Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), 4, durationOfBreak, onBreak, onReset);
+                .AdvancedCircuitBreaker(0.5, TimeSpan.FromSeconds(10), 4, durationOfBreak, onBreak, onReset);
 
             onResetCalled.Should().BeFalse();
         }
@@ -1597,7 +1591,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromSeconds(10),
                     minimumThroughput: 4,
@@ -1607,21 +1601,21 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
             // No adjustment to SystemClock.UtcNow, so all exceptions raised within same timeslice
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
 
@@ -1638,7 +1632,7 @@ namespace Polly.Specs
             var durationOfBreak = TimeSpan.FromSeconds(30);
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), 4, durationOfBreak, onBreak, onReset);
+                .AdvancedCircuitBreaker(0.5, TimeSpan.FromSeconds(10), 4, durationOfBreak, onBreak, onReset);
 
             onBreakCalled.Should().BeFalse();
 
@@ -1659,7 +1653,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromSeconds(10),
                     minimumThroughput: 4,
@@ -1669,31 +1663,31 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
             onBreakCalled.Should().Be(0);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
             onBreakCalled.Should().Be(0);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
             onBreakCalled.Should().Be(0);
 
             // No adjustment to SystemClock.UtcNow, so all exceptions raised within same timeslice
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
 
             onBreakCalled.Should().Be(1);
 
             // call through circuit when already broken - should not retrigger onBreak 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<BrokenCircuitException>();
 
             breaker.CircuitState.Should().Be(CircuitState.Open);
@@ -1701,7 +1695,7 @@ namespace Polly.Specs
         }
 
         [Fact]
-        public async Task Should_call_onreset_when_automatically_closing_circuit_but_not_when_halfopen()
+        public void Should_call_onreset_when_automatically_closing_circuit_but_not_when_halfopen()
         {
             int onBreakCalled = 0;
             int onResetCalled = 0;
@@ -1715,7 +1709,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromSeconds(10),
                     minimumThroughput: 4,
@@ -1725,21 +1719,21 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
             // No adjustment to SystemClock.UtcNow, so all exceptions raised within same timeslice
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
 
@@ -1753,13 +1747,13 @@ namespace Polly.Specs
             onResetCalled.Should().Be(0);
 
             // first call after duration is successful, so circuit should reset
-            await breaker.ExecuteAsync(() => Task.FromResult(true));
+            breaker.Execute(() => { });
             breaker.CircuitState.Should().Be(CircuitState.Closed);
             onResetCalled.Should().Be(1);
         }
 
         [Fact]
-        public async Task Should_not_call_onreset_on_successive_successful_calls()
+        public void Should_not_call_onreset_on_successive_successful_calls()
         {
             Action<Exception, TimeSpan> onBreak = (_, __) => { };
             bool onResetCalled = false;
@@ -1768,21 +1762,21 @@ namespace Polly.Specs
             var durationOfBreak = TimeSpan.FromSeconds(30);
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), 4, durationOfBreak, onBreak, onReset);
+                .AdvancedCircuitBreaker(0.5, TimeSpan.FromSeconds(10), 4, durationOfBreak, onBreak, onReset);
 
             onResetCalled.Should().BeFalse();
 
-            await breaker.ExecuteAsync(() => Task.FromResult(true));
+            breaker.Execute(() => { });
             breaker.CircuitState.Should().Be(CircuitState.Closed);
             onResetCalled.Should().BeFalse();
 
-            await breaker.ExecuteAsync(() => Task.FromResult(true));
+            breaker.Execute(() => { });
             breaker.CircuitState.Should().Be(CircuitState.Closed);
             onResetCalled.Should().BeFalse();
         }
 
         [Fact]
-        public async Task Should_call_onhalfopen_when_automatically_transitioning_to_halfopen_due_to_subsequent_execution()
+        public void Should_call_onhalfopen_when_automatically_transitioning_to_halfopen_due_to_subsequent_execution()
         {
             int onBreakCalled = 0;
             int onResetCalled = 0;
@@ -1798,7 +1792,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromSeconds(10),
                     minimumThroughput: 4,
@@ -1809,24 +1803,24 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
             onBreakCalled.Should().Be(0);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
             onBreakCalled.Should().Be(0);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
             onBreakCalled.Should().Be(0);
 
             // No adjustment to SystemClock.UtcNow, so all exceptions raised within same timeslice
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
 
@@ -1837,7 +1831,7 @@ namespace Polly.Specs
             onHalfOpenCalled.Should().Be(0); // not yet transitioned to half-open, because we have not queried state
 
             // first call after duration is successful, so circuit should reset
-            await breaker.ExecuteAsync(() => Task.FromResult(true));
+            breaker.Execute(() => { });
             onHalfOpenCalled.Should().Be(1); // called as action was placed for execution
             breaker.CircuitState.Should().Be(CircuitState.Closed);
             onResetCalled.Should().Be(1); // called after action succeeded
@@ -1860,7 +1854,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromSeconds(10),
                     minimumThroughput: 4,
@@ -1871,24 +1865,24 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
             onBreakCalled.Should().Be(0);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
             onBreakCalled.Should().Be(0);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
             onBreakCalled.Should().Be(0);
 
             // No adjustment to SystemClock.UtcNow, so all exceptions raised within same timeslice
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
 
@@ -1914,14 +1908,14 @@ namespace Polly.Specs
             var durationOfBreak = TimeSpan.FromSeconds(30);
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), 4, durationOfBreak, onBreak, onReset);
+                .AdvancedCircuitBreaker(0.5, TimeSpan.FromSeconds(10), 4, durationOfBreak, onBreak, onReset);
 
             onBreakCalled.Should().Be(0);
             breaker.Isolate();
             onBreakCalled.Should().Be(1);
 
             breaker.CircuitState.Should().Be(CircuitState.Isolated);
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
+            breaker.Invoking(x => x.Execute(() => { }))
                 .ShouldThrow<IsolatedCircuitException>();
 
             onResetCalled.Should().Be(0);
@@ -1929,8 +1923,7 @@ namespace Polly.Specs
             onResetCalled.Should().Be(1);
 
             breaker.CircuitState.Should().Be(CircuitState.Closed);
-            breaker.Awaiting(x => x.ExecuteAsync(() => Task.FromResult(true)))
-                .ShouldNotThrow();
+            breaker.Invoking(x => x.Execute(() => { })).ShouldNotThrow();
         }
 
         #region Tests that supplied context is passed to stage-change delegates
@@ -1948,7 +1941,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromSeconds(10),
                     minimumThroughput: 4,
@@ -1958,19 +1951,19 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>(
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>(
                 new { key1 = "value1", key2 = "value2" }.AsDictionary()
                 )).ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
@@ -1981,7 +1974,7 @@ namespace Polly.Specs
         }
 
         [Fact]
-        public async Task Should_call_onreset_with_the_passed_context()
+        public void Should_call_onreset_with_the_passed_context()
         {
             IDictionary<string, object> contextData = null;
 
@@ -1995,7 +1988,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromSeconds(10),
                     minimumThroughput: 4,
@@ -2005,19 +1998,19 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
 
             breaker.CircuitState.Should().Be(CircuitState.Open);
@@ -2027,7 +2020,7 @@ namespace Polly.Specs
 
 
             // first call after duration should invoke onReset, with context
-            await breaker.ExecuteAsync(() => Task.FromResult(true), new { key1 = "value1", key2 = "value2" }.AsDictionary());
+            breaker.Execute(() => { }, new { key1 = "value1", key2 = "value2" }.AsDictionary());
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
             contextData.Should()
@@ -2048,7 +2041,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromSeconds(10),
                     minimumThroughput: 4,
@@ -2058,19 +2051,19 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
 
             breaker.CircuitState.Should().Be(CircuitState.Open);
@@ -2079,7 +2072,7 @@ namespace Polly.Specs
         }
 
         [Fact]
-        public async Task Should_create_new_context_for_each_call_to_execute()
+        public void Should_create_new_context_for_each_call_to_execute()
         {
             string contextValue = null;
 
@@ -2095,7 +2088,7 @@ namespace Polly.Specs
 
             CircuitBreakerPolicy breaker = Policy
                 .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(
+                .AdvancedCircuitBreaker(
                     failureThreshold: 0.5,
                     timesliceDuration: TimeSpan.FromSeconds(10),
                     minimumThroughput: 4,
@@ -2105,19 +2098,19 @@ namespace Polly.Specs
                 );
 
             // Four of four actions in this test throw handled failures.
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>())
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Closed);
 
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>(new { key = "original_value" }.AsDictionary()))
+            breaker.Invoking(x => x.RaiseException<DivideByZeroException>(new { key = "original_value" }.AsDictionary()))
                 .ShouldThrow<DivideByZeroException>();
             breaker.CircuitState.Should().Be(CircuitState.Open);
             contextValue.Should().Be("original_value");
@@ -2129,289 +2122,12 @@ namespace Polly.Specs
             // but not yet reset
 
             // first call after duration is successful, so circuit should reset
-            await breaker.ExecuteAsync(() => Task.FromResult(true), new { key = "new_value" }.AsDictionary());
-
+            breaker.Execute(() => { }, new { key = "new_value" }.AsDictionary());
             breaker.CircuitState.Should().Be(CircuitState.Closed);
             contextValue.Should().Be("new_value");
         }
 
         #endregion
-
-        #region Cancellation support
-
-        [Fact]
-        public void Should_execute_action_when_non_faulting_and_cancellationtoken_not_cancelled()
-        {
-            var durationOfBreak = TimeSpan.FromMinutes(1);
-            CircuitBreakerPolicy breaker = Policy
-                .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), 4, durationOfBreak);
-
-            CancellationTokenSource cancellationTokenSource = new CancellationTokenSource();
-            CancellationToken cancellationToken = cancellationTokenSource.Token;
-
-            int attemptsInvoked = 0;
-            Action onExecute = () => attemptsInvoked++;
-
-            Scenario scenario = new Scenario
-            {
-                NumberOfTimesToRaiseException = 0,
-                AttemptDuringWhichToCancel = null,
-            };
-
-            breaker.Awaiting(x => x.RaiseExceptionAndOrCancellationAsync<DivideByZeroException>(scenario, cancellationTokenSource, onExecute))
-                .ShouldNotThrow();
-
-            attemptsInvoked.Should().Be(1);
-        }
-
-        [Fact]
-        public void Should_not_execute_action_when_cancellationtoken_cancelled_before_execute()
-        {
-            var durationOfBreak = TimeSpan.FromMinutes(1);
-            CircuitBreakerPolicy breaker = Policy
-                .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), 4, durationOfBreak);
-
-            CancellationTokenSource cancellationTokenSource = new CancellationTokenSource();
-            CancellationToken cancellationToken = cancellationTokenSource.Token;
-
-            int attemptsInvoked = 0;
-            Action onExecute = () => attemptsInvoked++;
-
-            Scenario scenario = new Scenario
-            {
-                NumberOfTimesToRaiseException = 0,
-                AttemptDuringWhichToCancel = null, // Cancellation token cancelled manually below - before any scenario execution.
-            };
-
-            cancellationTokenSource.Cancel();
-
-            breaker.Awaiting(x => x.RaiseExceptionAndOrCancellationAsync<DivideByZeroException>(scenario, cancellationTokenSource, onExecute))
-                .ShouldThrow<TaskCanceledException>()
-                .And.CancellationToken.Should().Be(cancellationToken);
-
-            attemptsInvoked.Should().Be(0);
-        }
-
-        [Fact]
-        public void Should_report_cancellation_during_otherwise_non_faulting_action_execution_when_user_delegate_observes_cancellationtoken()
-        {
-            var durationOfBreak = TimeSpan.FromMinutes(1);
-            CircuitBreakerPolicy breaker = Policy
-                .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), 4, durationOfBreak);
-
-            CancellationTokenSource cancellationTokenSource = new CancellationTokenSource();
-            CancellationToken cancellationToken = cancellationTokenSource.Token;
-
-            int attemptsInvoked = 0;
-            Action onExecute = () => attemptsInvoked++;
-
-            Scenario scenario = new Scenario
-            {
-                NumberOfTimesToRaiseException = 0,
-                AttemptDuringWhichToCancel = 1,
-                ActionObservesCancellation = true
-            };
-
-            breaker.Awaiting(x => x.RaiseExceptionAndOrCancellationAsync<DivideByZeroException>(scenario, cancellationTokenSource, onExecute))
-                .ShouldThrow<TaskCanceledException>()
-                .And.CancellationToken.Should().Be(cancellationToken);
-
-            attemptsInvoked.Should().Be(1);
-        }
-
-        [Fact]
-        public void Should_report_cancellation_during_faulting_action_execution_when_user_delegate_observes_cancellationtoken()
-        {
-            var durationOfBreak = TimeSpan.FromMinutes(1);
-            CircuitBreakerPolicy breaker = Policy
-                .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), 4, durationOfBreak);
-
-            CancellationTokenSource cancellationTokenSource = new CancellationTokenSource();
-            CancellationToken cancellationToken = cancellationTokenSource.Token;
-
-            int attemptsInvoked = 0;
-            Action onExecute = () => attemptsInvoked++;
-
-            Scenario scenario = new Scenario
-            {
-                NumberOfTimesToRaiseException = 1,
-                AttemptDuringWhichToCancel = 1,
-                ActionObservesCancellation = true
-            };
-
-            breaker.Awaiting(x => x.RaiseExceptionAndOrCancellationAsync<DivideByZeroException>(scenario, cancellationTokenSource, onExecute))
-                .ShouldThrow<TaskCanceledException>()
-                .And.CancellationToken.Should().Be(cancellationToken);
-
-            attemptsInvoked.Should().Be(1);
-        }
-
-        [Fact]
-        public void Should_report_cancellation_during_faulting_action_execution_when_user_delegate_does_not_observe_cancellationtoken()
-        {
-            var durationOfBreak = TimeSpan.FromMinutes(1);
-            CircuitBreakerPolicy breaker = Policy
-                .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), 4, durationOfBreak);
-
-            CancellationTokenSource cancellationTokenSource = new CancellationTokenSource();
-            CancellationToken cancellationToken = cancellationTokenSource.Token;
-
-            int attemptsInvoked = 0;
-            Action onExecute = () => attemptsInvoked++;
-
-            Scenario scenario = new Scenario
-            {
-                NumberOfTimesToRaiseException = 1,
-                AttemptDuringWhichToCancel = 1,
-                ActionObservesCancellation = false
-            };
-
-            breaker.Awaiting(x => x.RaiseExceptionAndOrCancellationAsync<DivideByZeroException>(scenario, cancellationTokenSource, onExecute))
-                .ShouldThrow<TaskCanceledException>()
-                .And.CancellationToken.Should().Be(cancellationToken);
-
-            attemptsInvoked.Should().Be(1);
-        }
-
-        [Fact]
-        public void Should_report_cancellation_when_both_open_circuit_and_cancellation()
-        {
-            var durationOfBreak = TimeSpan.FromMinutes(1);
-            CircuitBreakerPolicy breaker = Policy
-                .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), 1, durationOfBreak);
-
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
-                .ShouldThrow<DivideByZeroException>();
-
-            breaker.Awaiting(x => x.RaiseExceptionAsync<DivideByZeroException>())
-                .ShouldThrow<BrokenCircuitException>()
-                .WithMessage("The circuit is now open and is not allowing calls.")
-                .WithInnerException<DivideByZeroException>();
-            // Circuit is now broken.
-
-            CancellationTokenSource cancellationTokenSource = new CancellationTokenSource();
-            CancellationToken cancellationToken = cancellationTokenSource.Token;
-
-            int attemptsInvoked = 0;
-            Action onExecute = () => attemptsInvoked++;
-
-            cancellationTokenSource.Cancel();
-
-            Scenario scenario = new Scenario
-            {
-                NumberOfTimesToRaiseException = 1,
-                AttemptDuringWhichToCancel = null, // Cancelled manually instead - see above.
-                ActionObservesCancellation = false
-            };
-
-            breaker.Awaiting(x => x.RaiseExceptionAndOrCancellationAsync<DivideByZeroException>(scenario, cancellationTokenSource, onExecute))
-                .ShouldThrow<TaskCanceledException>()
-                .And.CancellationToken.Should().Be(cancellationToken);
-
-            attemptsInvoked.Should().Be(0);
-        }
-
-        [Fact]
-        public void Should_honour_different_cancellationtoken_captured_implicitly_by_action()
-        {
-            // Before CancellationToken support was built in to Polly, users of the library may have implicitly captured a CancellationToken and used it to cancel actions.  For backwards compatibility, Polly should not confuse these with its own CancellationToken; it should distinguish TaskCanceledExceptions thrown with different CancellationTokens.
-
-            var durationOfBreak = TimeSpan.FromMinutes(1);
-            CircuitBreakerPolicy breaker = Policy
-                .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), 4, durationOfBreak);
-
-            CancellationTokenSource policyCancellationTokenSource = new CancellationTokenSource();
-            CancellationToken policyCancellationToken = policyCancellationTokenSource.Token;
-
-            CancellationTokenSource implicitlyCapturedActionCancellationTokenSource = new CancellationTokenSource();
-            CancellationToken implicitlyCapturedActionCancellationToken = implicitlyCapturedActionCancellationTokenSource.Token;
-
-            implicitlyCapturedActionCancellationTokenSource.Cancel();
-
-            int attemptsInvoked = 0;
-
-            breaker.Awaiting(x => x.ExecuteAsync(async ct =>
-            {
-                attemptsInvoked++;
-                await Task.FromResult(true);
-                implicitlyCapturedActionCancellationToken.ThrowIfCancellationRequested();
-            }, policyCancellationToken))
-                .ShouldThrow<TaskCanceledException>()
-                .And.CancellationToken.Should().Be(implicitlyCapturedActionCancellationToken);
-
-            attemptsInvoked.Should().Be(1);
-        }
-
-        [Fact]
-        public void Should_execute_func_returning_value_when_cancellationtoken_not_cancelled()
-        {
-            var durationOfBreak = TimeSpan.FromMinutes(1);
-            CircuitBreakerPolicy breaker = Policy
-                .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), 4, durationOfBreak);
-
-            CancellationTokenSource cancellationTokenSource = new CancellationTokenSource();
-            CancellationToken cancellationToken = cancellationTokenSource.Token;
-
-            int attemptsInvoked = 0;
-            Action onExecute = () => attemptsInvoked++;
-
-            bool? result = null;
-
-            Scenario scenario = new Scenario
-            {
-                NumberOfTimesToRaiseException = 0,
-                AttemptDuringWhichToCancel = null,
-            };
-
-            breaker.Awaiting(async x => result = await x.RaiseExceptionAndOrCancellationAsync<DivideByZeroException, bool>(scenario, cancellationTokenSource, onExecute, true).ConfigureAwait(false))
-                .ShouldNotThrow();
-
-            result.Should().BeTrue();
-
-            attemptsInvoked.Should().Be(1);
-        }
-
-        [Fact]
-        public void Should_honour_and_report_cancellation_during_func_execution()
-        {
-            var durationOfBreak = TimeSpan.FromMinutes(1);
-            CircuitBreakerPolicy breaker = Policy
-                .Handle<DivideByZeroException>()
-                .TimesliceCircuitBreakerAsync(0.5, TimeSpan.FromSeconds(10), 4, durationOfBreak);
-
-            CancellationTokenSource cancellationTokenSource = new CancellationTokenSource();
-            CancellationToken cancellationToken = cancellationTokenSource.Token;
-
-            int attemptsInvoked = 0;
-            Action onExecute = () => attemptsInvoked++;
-
-            bool? result = null;
-
-            Scenario scenario = new Scenario
-            {
-                NumberOfTimesToRaiseException = 0,
-                AttemptDuringWhichToCancel = 1,
-                ActionObservesCancellation = true
-            };
-
-            breaker.Awaiting(async x => result = await x.RaiseExceptionAndOrCancellationAsync<DivideByZeroException, bool>(scenario, cancellationTokenSource, onExecute, true).ConfigureAwait(false))
-                .ShouldThrow<TaskCanceledException>().And.CancellationToken.Should().Be(cancellationToken);
-
-            result.Should().Be(null);
-
-            attemptsInvoked.Should().Be(1);
-        }
-
-        #endregion
-
 
         #endregion
